+++ gcherry/tests/test_orbit_targeting.py
@@ -5,7 +5,7 @@ import openmdao.api as om
 from gcherry.guidance_components import (
     OrbitGuidanceGroup)
 from gcherry.transform import global2perifocal_rot
-from gcherry.log_utils_refactor import almost_equal
+from gcherry.log_utils import almost_equal
 
 
 # Takeoff from lunar surface along equator to a position with 0 r_dot.
