+++ gcherry/tests/test_outer_loop.py
@@ -4,7 +4,7 @@ import openmdao.api as om
 
 from gcherry.guidance_components import (
     OuterLoopGroupRefactor)
-from gcherry.log_utils_refactor import almost_equal
+from gcherry.log_utils import almost_equal
 
 
 # Takeoff from lunar surface along equator to a position with 0 r_dot.
